requirements.txt
alembic==1.0.0
aliyun-python-sdk-core==2.13.4
aliyun-python-sdk-core-v3==2.9.1
aliyun-python-sdk-ecs==4.16.3
aliyun-python-sdk-kms==2.5.0
amqp==2.6.1
aniso8601==3.0.2
appnope==0.1.0
APScheduler==3.5.3
argcomplete==1.8.1
asn1crypto==0.24.0
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.6.0
behave==1.2.6
billiard==3.6.3.0
bitsets==0.7.14
celery==4.4.7
certifi==2018.8.24
cffi==1.14.3
chardet==3.0.4
cheap-repr==0.4.4
click==6.7
colorama==0.4.1
concepts==0.8
cos-python-sdk-v5==1.6.7
crayons==0.3.1
crcmod==1.7
cryptography==2.3.1
decorator==4.4.2
demjson==2.2.4
dicttoxml==1.7.4
distro==1.4.0
dnspython==1.16.0
docopt==0.6.2
dominate==2.3.5
elasticsearch==6.3.1
eventlet==0.24.1
fileconfig==0.5.8
Flask==1.0.2
flask-apidoc==1.1.2
Flask-APScheduler==1.10.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.7
Flask-Docs==0.2.2
Flask-JWT==0.3.2
Flask-Migrate==2.2.1
Flask-request-params==0.3.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
get-port==0.0.5
gevent==1.4.0
graphviz==0.10.1
greenlet==0.4.15
grpcio==1.32.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.7
import-string==0.1.0
importlib-metadata==2.0.0
ipython==7.18.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.17.2
Jinja2==2.10
jmespath==0.9.4
jsonpickle==1.4.1
keyring==10.6.0
keyrings.alt==3.0
kombu==4.6.11
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lxml==4.2.1
Mako==1.0.7
manage==0.1.13
MarkupSafe==1.1.1
monotonic==1.5
more-itertools==8.5.0
numpy==1.19.2
oauth==1.0.1
oauthlib==3.0.1
olefile==0.45
oss2==2.5.0
pandas==1.1.2
parse==1.12.0
parse-type==0.4.2
parso==0.7.1
pbr==5.2.0
pexpect==4.2.1
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
pycparser==2.20
pycryptodome==3.9.8
Pygments==2.7.1
PyJWT==1.4.2
PyMySQL==0.10.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
qcloudsms-py==0.1.4
redis==3.5.3
removebg==0.4
requests==2.24.0
shortuuid==1.0.1
six==1.15.0
SQLAlchemy==1.3.20
testresources==2.0.1
traitlets==5.0.4
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.10
vine==1.3.0
visitor==0.1.3
wadllib==1.3.4
wcwidth==0.2.5
webencodings==0.5.1
websocket==0.2.1
websocket-client==0.57.0
Werkzeug==1.0.1
xlwt==1.3.0
zipp==3.3.0
zope.event==4.5.0
zope.interface==5.1.2
